refactor: statistics 패키지 API를 report 패키지로 통합
- getWeeklyStatistics, getMissionTypeStatistics 중복 API 제거
- getMonthlyPattern을 report 패키지로 이동
- MonthlyPatternResponse DTO를 report.dto로 이동
- StatisticsController, StatisticsService 제거
- WeeklyReportController에 월간 패턴 분석 API 추가
- 불필요한 DTO 및 테스트 파일 정리
